move OverlayStyle to /overlayStyle
