+++ src/modules/olMap/overlayStyle/OverlayStyle.js
